ResizeMatrix

Resize a matrix in new dimension

ResizeMatrix(
  mat.mtx = NULL,
  rowDim.num = NULL,
  colDim.num = NULL,
  rowKeep.ndx = NULL,
  colKeep.ndx = NULL
)

Arguments

mat.mtx

<matrix>: a Matrix to resize

rowDim.num

<integer>: the number of rows in reiszed matrix

colDim.num

<integer>: the number of columns in reiszed matrix

rowKeep.ndx

<integer>: index of one row to keep in resized matrix

colKeep.ndx

<integer>: index of one column to keep in resized matrix

Value

resized matrix

Examples

mat.mtx <- matrix(0,11,11)
mat.mtx[which(as.logical(1:(11*11)%%2))] <- 1:ceiling((11*11)/2)
mat.mtx[2,] <- 100
mat.mtx[,7] <- 200
mat.mtx
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11]
#>  [1,]    1    0   12    0   23    0  200    0   45     0    56
#>  [2,]  100  100  100  100  100  100  200  100  100   100   100
#>  [3,]    2    0   13    0   24    0  200    0   46     0    57
#>  [4,]    0    8    0   19    0   30  200   41    0    52     0
#>  [5,]    3    0   14    0   25    0  200    0   47     0    58
#>  [6,]    0    9    0   20    0   31  200   42    0    53     0
#>  [7,]    4    0   15    0   26    0  200    0   48     0    59
#>  [8,]    0   10    0   21    0   32  200   43    0    54     0
#>  [9,]    5    0   16    0   27    0  200    0   49     0    60
#> [10,]    0   11    0   22    0   33  200   44    0    55     0
#> [11,]    6    0   17    0   28    0  200    0   50     0    61
ResizeMatrix(mat.mtx=mat.mtx, rowDim.num=7, colDim.num=7, rowKeep.ndx=NULL, colKeep.ndx=NULL)
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7]
#> [1,]    1   12    0    0    0   45   56
#> [2,]    2   13    0    0    0   46   57
#> [3,]    0    0   19   30   41    0    0
#> [4,]    0    0   20   31   42    0    0
#> [5,]    0    0   21   32   43    0    0
#> [6,]    5   16    0    0    0   49   60
#> [7,]    6   17    0    0    0   50   61
ResizeMatrix(mat.mtx=mat.mtx, rowDim.num=7, colDim.num=7, rowKeep.ndx=2, colKeep.ndx=7)
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7]
#> [1,]  100  100  100  200  100  100  100
#> [2,]    2   13    0  200    0   46   57
#> [3,]    0    0   19  200   41    0    0
#> [4,]    0    0   20  200   42    0    0
#> [5,]    0    0   21  200   43    0    0
#> [6,]    5   16    0  200    0   49   60
#> [7,]    6   17    0  200    0   50   61
ResizeMatrix(mat.mtx=mat.mtx, rowDim.num=13, colDim.num=13, rowKeep.ndx=NULL, colKeep.ndx=NULL)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    1    0   12    0    0   23    0  200    0     0    45     0    56
#>  [2,]  100  100  100  100  100  100  100  200  100   100   100   100   100
#>  [3,]    2    0   13    0    0   24    0  200    0     0    46     0    57
#>  [4,]    0    8    0   19   19    0   30  200   41    41     0    52     0
#>  [5,]    0    8    0   19   19    0   30  200   41    41     0    52     0
#>  [6,]    3    0   14    0    0   25    0  200    0     0    47     0    58
#>  [7,]    0    9    0   20   20    0   31  200   42    42     0    53     0
#>  [8,]    4    0   15    0    0   26    0  200    0     0    48     0    59
#>  [9,]    0   10    0   21   21    0   32  200   43    43     0    54     0
#> [10,]    0   10    0   21   21    0   32  200   43    43     0    54     0
#> [11,]    5    0   16    0    0   27    0  200    0     0    49     0    60
#> [12,]    0   11    0   22   22    0   33  200   44    44     0    55     0
#> [13,]    6    0   17    0    0   28    0  200    0     0    50     0    61
ResizeMatrix(mat.mtx=mat.mtx, rowDim.num=13, colDim.num=13, rowKeep.ndx=2, colKeep.ndx=7)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    1    0   12    0    0   23    0  200    0     0    45     0    56
#>  [2,]  100  100  100  100  100  100  100  200  100   100   100   100   100
#>  [3,]    2    0   13    0    0   24    0  200    0     0    46     0    57
#>  [4,]    0    8    0   19   19    0   30  200   41    41     0    52     0
#>  [5,]    0    8    0   19   19    0   30  200   41    41     0    52     0
#>  [6,]    3    0   14    0    0   25    0  200    0     0    47     0    58
#>  [7,]    0    9    0   20   20    0   31  200   42    42     0    53     0
#>  [8,]    4    0   15    0    0   26    0  200    0     0    48     0    59
#>  [9,]    0   10    0   21   21    0   32  200   43    43     0    54     0
#> [10,]    0   10    0   21   21    0   32  200   43    43     0    54     0
#> [11,]    5    0   16    0    0   27    0  200    0     0    49     0    60
#> [12,]    0   11    0   22   22    0   33  200   44    44     0    55     0
#> [13,]    6    0   17    0    0   28    0  200    0     0    50     0    61